passe les injections dans les fichiers respectifs
