+++ journal.js
@@ -1,4 +1,4 @@
-import { convertToNumber, trouverCompte, sommeCompteParRacine } from './utils.js';
+import { convertToNumber, trouverCompte, sommeCompteParRacine, formatToCurrency } from './utils.js';
 
 function displayErrorMessage(message) {
     const errorMessageElement = document.getElementById('error-message');
@@ -153,4 +153,17 @@ export function arretComptesClotureEcritures(ecritures, currentYear) {
     }
 
     return ecrituresArret;
+}
+
+export function injecteJournalEcritures(journalEcritures) {
+    const tableBody = document.getElementById('journal-ecritures');
+    tableBody.innerHTML = journalEcritures.map(ecriture => `
+        <tr>
+            <td>${ecriture['Date']}</td>
+            <td>${ecriture['Compte']}</td>
+            <td>${ecriture['Libellé']}</td>
+            <td>${formatToCurrency(ecriture['Débit (€)'])}</td>
+            <td>${formatToCurrency(ecriture['Crédit (€)'])}</td>
+        </tr>
+    `).join('');
 }
